move all the helper functions to the helper module
